[PF-127] Don't rollback creation flight if resource is READY (#18)
* update

* split steps

* split steps

* split steps

* add test dir

* add test dir

* add test dir

* add test dir

* add test dir

* address comment

* address comment

* delete flightManager test

* tmp

* add one more test

* add one more test

* address comment

* address comment

* address comments

* address comments

* more polling times since flight make take long

* more polling times since flight make take long

* move enums/autovalue classes to common

* don

* fix test

* fix test

* fix test

* address comment

* address comment

* address comment

* address comment

* address comment

+++ src/main/java/bio/terra/rbs/service/resource/flight/CreateResourceDbEntityStep.java
@@ -1,5 +1,7 @@
 package bio.terra.rbs.service.resource.flight;
 
+import static bio.terra.rbs.service.resource.flight.StepUtils.isResourceReady;
+
 import bio.terra.rbs.common.PoolId;
 import bio.terra.rbs.common.Resource;
 import bio.terra.rbs.common.ResourceId;
@@ -39,6 +41,9 @@ public class CreateResourceDbEntityStep implements Step {
 
   @Override
   public StepResult undoStep(FlightContext flightContext) {
+    if (isResourceReady(flightContext)) {
+      return StepResult.getStepResultSuccess();
+    }
     // Just delete the resource entity if creation not succeed. There is no need to keep this
     // record.
     rbsDao.deleteResource(ResourceId.retrieve(flightContext.getWorkingMap()));
